L1-T1 is fixed
